[DFM-1001] Refactor tests
